refractor: move reason list ot zustand store
