+++ src/components/Appbar.tsx
@@ -14,23 +14,21 @@ import { listenTo } from '../utils/window';
 import FilterMenu from './FilterMenu';
 import { filterTransactionAtom, writeFilterTransactionAtom } from '../store/transaction';
 import CloseButton from './CloseButton';
-import { fetchReasonsAtom, reasonsAtom, reasonsMapAtom } from '../store/reason';
 import MagnifyIcon from './icons/Magnify';
 import PlusOneIcon from './icons/PlusOne';
 import { Maybe } from '../graphql.generated';
 import ExitIcon from './icons/Exit';
 import { Button } from './Button';
 import { remove } from '../utils/auth';
-import selectedProvider from '../store/provider';
+import selectedProvider from '../dataSource';
 import { useAuthStore } from '../store/auth';
+import { useReasonStore } from '../store/reason';
 
 export default function Appbar() {
   const navigate = useNavigate();
 
   const filtering = useAtomValue(filterTransactionAtom);
   const setFiltering = useUpdateAtom(writeFilterTransactionAtom);
-  const fetchReason = useUpdateAtom(fetchReasonsAtom);
-  const reasonList = useAtomValue(reasonsAtom);
 
   const { setAuth } = useAuthStore();
 
@@ -69,12 +67,6 @@ export default function Appbar() {
   const handleClose = useCallback(() => {
     setOpenFilter(false);
   }, []);
-
-  useEffect(() => {
-    if (!reasonList.length) {
-      fetchReason();
-    }
-  }, [reasonList]);
 
   useEffect(
     () =>
@@ -224,14 +216,14 @@ function ReasonChip({
   reasons?: Maybe<number[]>;
   onDelete?: (reasonId: number) => void;
 }) {
-  const reasonsMap = useAtomValue(reasonsMapAtom);
+  const { reasonsMap } = useReasonStore();
 
   if (!reasons?.length) return null;
 
   return (
     <>
       {reasons?.map((r) => {
-        const reason = reasonsMap[r];
+        const reason = reasonsMap?.get(r);
         if (!reason) return null;
 
         return (
